move to using uint32_t for vertex_t and edge_t, better performance on atomic adds and shouldnt need the 4billion range

+++ src/dist_bfs_kernels.cuh
@@ -4,8 +4,8 @@
 #include <cuda/experimental/stf.cuh>
 
 using namespace cuda::experimental::stf;
-using vertex_t = int64_t;
-using edge_t = int64_t;
+using vertex_t = uint32_t;
+using edge_t = uint32_t;
 
 // Reset the next_frontier_size counter
 __global__ void reset_counter_kernel(slice<vertex_t> counter) {
